chore: improve Stat exports
